Moved the components of the manage course class back into the same file because behaviors for buttons will likely need to reference the course name stored in that menu
